refactor: move AppError from common-meta-types to common-meta-app; move KVAppError from common-meta-types to common-meta-api
`AppError` and its sub errors are meta-service application level errors,
such as `UnknownDatabaseId`. Move it to the crate that defines
meta application data types: `common-meta-app`.

`KVAppError` is a super set of `AppError` and `MetaError`, is used by
`KVApi` based API implementation,  such as `SchemaApi`, to return an
error from either meta-service or application implementation.
