+++ scripts/ci/ci-run-sqllogic-tests.sh
