+++ src/meta/api/Cargo.toml
@@ -17,6 +17,7 @@ common-exception = { path = "../../common/exception" }
 common-expression = { path = "../../query/expression" }
 common-meta-app = { path = "../app" }
 common-meta-kvapi = { path = "../kvapi" }
+common-meta-stoerr = { path = "../stoerr" }
 common-meta-types = { path = "../types" }
 common-proto-conv = { path = "../proto-conv" }
 common-protos = { path = "../protos" }
