+++ src/meta/api/src/kv_app_error.rs
@@ -1,29 +1,28 @@
-// Copyright 2022 Datafuse Labs.
+//  Copyright 2023 Datafuse Labs.
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
 
 use std::any::type_name;
 
 use common_exception::ErrorCode;
+use common_meta_app::app_error::AppError;
 use common_meta_stoerr::MetaStorageError;
+use common_meta_types::InvalidReply;
+use common_meta_types::MetaAPIError;
+use common_meta_types::MetaClientError;
+use common_meta_types::MetaError;
+use common_meta_types::MetaNetworkError;
 use tonic::Status;
-
-use crate::AppError;
-use crate::InvalidReply;
-use crate::MetaAPIError;
-use crate::MetaClientError;
-use crate::MetaError;
-use crate::MetaNetworkError;
 
 /// Errors for a kvapi::KVApi based application, such SchemaApi, ShareApi.
 ///
