+++ src/meta/api/src/lib.rs
@@ -16,9 +16,10 @@
 #![deny(unused_crate_dependencies)]
 extern crate common_meta_types;
 
-pub mod errors;
+pub mod compat_errors;
 mod id;
 mod id_generator;
+pub mod kv_app_error;
 pub mod reply;
 mod schema_api;
 mod schema_api_impl;
