+++ src/meta/api/src/schema_api_impl.rs
@@ -19,6 +19,28 @@ use std::sync::Arc;
 
 use chrono::DateTime;
 use chrono::Utc;
+use common_meta_app::app_error::AppError;
+use common_meta_app::app_error::CreateDatabaseWithDropTime;
+use common_meta_app::app_error::CreateTableWithDropTime;
+use common_meta_app::app_error::DatabaseAlreadyExists;
+use common_meta_app::app_error::DropDbWithDropTime;
+use common_meta_app::app_error::DropTableWithDropTime;
+use common_meta_app::app_error::ShareHasNoGrantedDatabase;
+use common_meta_app::app_error::ShareHasNoGrantedPrivilege;
+use common_meta_app::app_error::TableAlreadyExists;
+use common_meta_app::app_error::TableVersionMismatched;
+use common_meta_app::app_error::TxnRetryMaxTimes;
+use common_meta_app::app_error::UndropDbHasNoHistory;
+use common_meta_app::app_error::UndropDbWithNoDropTime;
+use common_meta_app::app_error::UndropTableAlreadyExists;
+use common_meta_app::app_error::UndropTableHasNoHistory;
+use common_meta_app::app_error::UndropTableWithNoDropTime;
+use common_meta_app::app_error::UnknownDatabaseId;
+use common_meta_app::app_error::UnknownShareAccounts;
+use common_meta_app::app_error::UnknownTable;
+use common_meta_app::app_error::UnknownTableId;
+use common_meta_app::app_error::WrongShare;
+use common_meta_app::app_error::WrongShareObject;
 use common_meta_app::schema::CountTablesKey;
 use common_meta_app::schema::CountTablesReply;
 use common_meta_app::schema::CountTablesReq;
@@ -78,32 +100,10 @@ use common_meta_app::share::ShareId;
 use common_meta_app::share::ShareNameIdent;
 use common_meta_kvapi::kvapi;
 use common_meta_kvapi::kvapi::Key;
-use common_meta_types::errors::app_error::AppError;
-use common_meta_types::errors::app_error::CreateDatabaseWithDropTime;
-use common_meta_types::errors::app_error::CreateTableWithDropTime;
-use common_meta_types::errors::app_error::DatabaseAlreadyExists;
-use common_meta_types::errors::app_error::DropDbWithDropTime;
-use common_meta_types::errors::app_error::DropTableWithDropTime;
-use common_meta_types::errors::app_error::ShareHasNoGrantedDatabase;
-use common_meta_types::errors::app_error::ShareHasNoGrantedPrivilege;
-use common_meta_types::errors::app_error::TableAlreadyExists;
-use common_meta_types::errors::app_error::TableVersionMismatched;
-use common_meta_types::errors::app_error::TxnRetryMaxTimes;
-use common_meta_types::errors::app_error::UndropDbHasNoHistory;
-use common_meta_types::errors::app_error::UndropDbWithNoDropTime;
-use common_meta_types::errors::app_error::UndropTableAlreadyExists;
-use common_meta_types::errors::app_error::UndropTableHasNoHistory;
-use common_meta_types::errors::app_error::UndropTableWithNoDropTime;
-use common_meta_types::errors::app_error::UnknownShareAccounts;
-use common_meta_types::errors::app_error::UnknownTable;
-use common_meta_types::errors::app_error::UnknownTableId;
-use common_meta_types::errors::app_error::WrongShare;
-use common_meta_types::errors::app_error::WrongShareObject;
 use common_meta_types::ConditionResult;
 use common_meta_types::GCDroppedDataReply;
 use common_meta_types::GCDroppedDataReq;
 use common_meta_types::InvalidReply;
-use common_meta_types::KVAppError;
 use common_meta_types::MatchSeqExt;
 use common_meta_types::MetaError;
 use common_meta_types::MetaId;
@@ -111,7 +111,6 @@ use common_meta_types::MetaNetworkError;
 use common_meta_types::TxnCondition;
 use common_meta_types::TxnOp;
 use common_meta_types::TxnRequest;
-use common_meta_types::UnknownDatabaseId;
 use common_tracing::func_name;
 use tracing::debug;
 use tracing::error;
@@ -125,6 +124,7 @@ use crate::get_share_or_err;
 use crate::get_struct_value;
 use crate::get_u64_value;
 use crate::is_db_need_to_be_remove;
+use crate::kv_app_error::KVAppError;
 use crate::list_keys;
 use crate::list_u64_value;
 use crate::send_txn;
