+++ src/meta/api/src/share_api_test_suite.rs
@@ -22,7 +22,6 @@ use common_meta_app::schema::TableMeta;
 use common_meta_app::schema::TableNameIdent;
 use common_meta_app::share::*;
 use common_meta_kvapi::kvapi;
-use common_meta_types::KVAppError;
 use common_meta_types::MetaError;
 use enumflags2::BitFlags;
 use tracing::info;
@@ -33,6 +32,7 @@ use crate::get_share_id_to_name_or_err;
 use crate::get_share_meta_by_id_or_err;
 use crate::get_share_or_err;
 use crate::is_all_db_data_removed;
+use crate::kv_app_error::KVAppError;
 use crate::testing::get_kv_data;
 use crate::SchemaApi;
 use crate::ShareApi;
