+++ src/meta/api/src/testing.rs
@@ -16,12 +16,13 @@
 
 use common_meta_kvapi::kvapi;
 use common_meta_types::anyerror::AnyError;
-use common_meta_types::KVAppError;
 use common_meta_types::MetaAPIError;
 use common_meta_types::MetaDataError;
 use common_meta_types::MetaDataReadError;
 use common_meta_types::MetaError;
 use common_proto_conv::FromToProto;
+
+use crate::kv_app_error::KVAppError;
 
 /// Get existing value by key. Panic if key is absent.
 pub(crate) async fn get_kv_data<T>(
