+++ src/meta/app/src/lib.rs
@@ -22,6 +22,7 @@
 #![deny(unused_crate_dependencies)]
 #![feature(no_sanitize)]
 
+pub mod app_error;
 pub mod principal;
 pub mod schema;
 pub mod share;
