+++ src/meta/client/Cargo.toml
@@ -31,7 +31,6 @@ prost = { workspace = true }
 semver = "1.0.14"
 serde = { workspace = true }
 serde_json = { workspace = true }
-thiserror = { workspace = true }
 tonic = { version = "0.8.1", features = ["transport", "codegen", "prost", "tls-roots", "tls"] }
 tracing = "0.1.36"
