+++ src/meta/types/src/lib.rs
@@ -54,27 +54,6 @@ pub use cluster::NodeInfo;
 pub use cmd::Cmd;
 pub use cmd::UpsertKV;
 pub use endpoint::Endpoint;
-pub use errors::app_error::AppError;
-pub use errors::app_error::CreateDatabaseWithDropTime;
-pub use errors::app_error::CreateTableWithDropTime;
-pub use errors::app_error::DatabaseAlreadyExists;
-pub use errors::app_error::DropDbWithDropTime;
-pub use errors::app_error::DropTableWithDropTime;
-pub use errors::app_error::ShareAlreadyExists;
-pub use errors::app_error::TableAlreadyExists;
-pub use errors::app_error::TableVersionMismatched;
-pub use errors::app_error::UndropDbHasNoHistory;
-pub use errors::app_error::UndropDbWithNoDropTime;
-pub use errors::app_error::UndropTableAlreadyExists;
-pub use errors::app_error::UndropTableHasNoHistory;
-pub use errors::app_error::UndropTableWithNoDropTime;
-pub use errors::app_error::UnknownDatabase;
-pub use errors::app_error::UnknownDatabaseId;
-pub use errors::app_error::UnknownShare;
-pub use errors::app_error::UnknownTable;
-pub use errors::app_error::UnknownTableId;
-pub use errors::app_error::WrongShareObject;
-pub use errors::kv_app_errors::KVAppError;
 pub use errors::meta_api_errors::MetaAPIError;
 pub use errors::meta_api_errors::MetaDataError;
 pub use errors::meta_api_errors::MetaDataReadError;
