+++ src/query/management/src/stage/stage_mgr.rs
@@ -21,11 +21,11 @@ use common_meta_api::reply::txn_reply_to_api_result;
 use common_meta_api::txn_cond_seq;
 use common_meta_api::txn_op_del;
 use common_meta_api::txn_op_put;
+use common_meta_app::app_error::TxnRetryMaxTimes;
 use common_meta_app::principal::StageFile;
 use common_meta_app::principal::UserStageInfo;
 use common_meta_kvapi::kvapi;
 use common_meta_kvapi::kvapi::UpsertKVReq;
-use common_meta_types::errors::app_error::TxnRetryMaxTimes;
 use common_meta_types::ConditionResult::Eq;
 use common_meta_types::MatchSeq;
 use common_meta_types::MatchSeqExt;
